pyproject.toml
[tool.poetry]
name = "planeks_test_task"
version = "0.1.0"
description = "tesk task for planeks.com"
authors = ["Vladyslav Babych"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.0"
psycopg2 = "^2.8.4"
django-crispy-forms = "^1.8.1"
requests = "^2.22.0"
django-bootstrap-datepicker-plus = "^3.0.5"
django-bootstrap4 = "^1.1.1"
django-ckeditor = "^5.8.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
